// website-frontend/lib/types.ts
export interface Value {
    iconPath: string;
    title: string;
    description: string;
}

export interface Stat {
    number: string;
    label: string;
    color: string;
}

export interface SuccessStat {
    value: string;
    label: string;
}

export interface Milestone {
    year: string;
    title: string;
    description: string;
}